Increase the resolution of MIDI pitch wheel events.
Previously we were only using the upper 7 bits of the available 14 bits. This enables much smoother pitch wheel behaviour, but required some changes to avoid generating thousands of MIDI events.

Bug: #54

+++ source/audio/midioutputdevice.h
@@ -53,7 +53,7 @@ public:
     bool setPatch(int channel, uint8_t patch);
     bool setVolume(int channel, uint8_t volume);
     bool setPan(int channel, uint8_t pan);
-    bool setPitchBend(int channel, uint8_t bend);
+    bool setPitchBend(int channel, uint16_t bend);
     bool playNote(int channel, uint8_t pitch, uint8_t velocity);
     bool stopNote(int channel, uint8_t pitch);
     bool setVibrato(int channel, uint8_t modulation);
